\name{heatmapGen-package}
\alias{heatmapGen-package}
\alias{heatmapGen}
\docType{package}
\title{
Heatmap Tools
}
\description{
This package is heavily based on the 'hmutil' package and the 'heatmap.n' framework originally developed by Henri Sara.
}
\details{
\tabular{ll}{
Package: \tab heatmapGen\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
License: \tab MIT License\cr
}
The key function in the package for end users is \code{heatmap.n} and there are several internal functions.

}
\author{
Author: Vidal Fey <vidal.fey@utu.fi> ; Maintainer: Dhanaprakash Jambulingam <dhajam@utu.fi>
}
\keyword{ package }
